Rename usecase to application, route to presentation
Align directory names with layer names for clarity.
